# Improve JSON str encode performance