This improves the performance of encoding JSON strings, mostly by rearranging some work and unrolling some loops.